fixed some issues

+++ controllers/Dashboard/Brand/BrandControllers.js
@@ -91,40 +91,61 @@ const Read = async (req, res) => {
 }
 
 const Update = async (req, res) => {
-    const updateStateQuery = `UPDATE brands SET name=? WHERE id=${req.body.id}`;
-    connection.query(updateStateQuery, [req.body?.name], (error, results) => {
-        if (error) {
-            res.status(500).send({
-                error: true,
-                data: [error],
-                message: `Something is wrong!`
+    if (req.body.id && req.body.name) {
+        const data = req.body.logo ? [req.body?.name, req.body?.logo] : [req.body?.name]
+        const updateBrandQuery = req.body.logo ?
+            `UPDATE brands SET name=? , logo=?  WHERE id=${req.body.id}`
+            :
+            `UPDATE brands SET name=?  WHERE id=${req.body.id}`;
+
+        connection.query(updateBrandQuery, data, (error, results) => {
+            if (error) {
+                res.status(500).send({
+                    error: true,
+                    data: [error],
+                    message: `Something is wrong!`
+                })
+            }
+            res.status(200).send({
+                error: false,
+                data: results,
+                message: 'Brand updated successfully.'
             })
-        }
-        res.status(200).send({
-            error: false,
-            data: results,
-            message: 'State updated successfully.'
         })
-    })
+    } else {
+        res.status(500).send({
+            error: true,
+            data: [],
+            message: `name and id are required.`
+        })
+    }
 }
 
 const Delete = async (req, res) => {
-    const deleteSingleStateQuery = `DELETE FROM brands WHERE id=${req.body.id}`;
+    if (req.body.id) {
+        const deleteSingleStateQuery = `DELETE FROM brands WHERE id=${req.body.id}`;
 
-    connection.query(deleteSingleStateQuery, (error, result) => {
-        if (error) {
-            res.status(500).send({
-                error: true,
-                data: [error],
-                message: `Something is wrong!`
+        connection.query(deleteSingleStateQuery, (error, result) => {
+            if (error) {
+                res.status(500).send({
+                    error: true,
+                    data: [error],
+                    message: `Something is wrong!`
+                })
+            }
+            res.status(200).send({
+                error: false,
+                data: result,
+                message: 'Deleted successfully.'
             })
-        }
-        res.status(200).send({
-            error: false,
-            data: result,
-            message: 'Deleted successfully.'
         })
-    })
+    }else{
+        res.status(500).send({
+            error: true,
+            data: [error],
+            message: `id is required!`
+        })
+    }
 }
 
 module.exports = {
